Add support for Dual Sim F5122
Tested on my Xperia X Dual and it works. This file is identical with F5121_34.2.A.0.266.fsc.
Although, in my testings, for F5122 and F5121, the related fsc should be renamed as "default.fsc" to be found by Flashtool, or it will fail. Another user using F5121 reported the same issue. Not sure what went wrong.